% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saliency_heatmap.R
\name{saliency_heatmap}
\alias{saliency_heatmap}
\title{Produce Salience heatmap from Raw MSI Net output}
\usage{
saliency_heatmap(
  saliency_files_directory,
  raw_image_directory,
  output_directory
)
}
\arguments{
\item{output_directory}{}
}
\value{
jpeg where heatmap is overlayed origninal creative
}
\description{
Produce Salience heatmap from Raw MSI Net output
}
